% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustCausalForest.R
\name{robustCausalForest}
\alias{robustCausalForest}
\title{Estimate Heterogeneous Treatment Effect via Causal Forest with Median Splitting Rules}
\usage{
robustCausalForest(
  Y,
  W,
  X,
  Y.hat,
  W.hat,
  Y.res,
  W.res,
  k,
  ntrees,
  split.R = "CT",
  cv.opt = "CT",
  ...
)
}
\arguments{
\item{Y}{outcome variable}

\item{W}{treatment variable}

\item{X}{covariate matrix}

\item{Y.hat}{cross-fitted estimates of E[Y|X]}

\item{W.hat}{cross-fitted estimates of E[W|X]}

\item{Y.res}{outcome residuals}

\item{W.res}{treatment residuals}

\item{k}{number of folds for cross-fitting}

\item{ntrees}{number of trees in forest}

\item{split.R}{splitting rule (one of "CT", "MSD", "MAD", "LMS")}

\item{cv.opt}{cross-valdiation option (one of "CT", "MSD", "MAD", "LMS")}
}
\value{
A list with two elements. The first one is \code{pred_cate} which holds
the predicted CATE and the corresponding standard error for each unit.
The second element is \code{pred_ate} which holds the ATE and its corresponding
standard error, calculated based on debiased CATEs through an AIPW estimator.
}
\description{
Estimate heterogeneous treatment effect based on random forests.
}
